* evandorsky/joy:
  Default mode free
  Fix my scons file and change default state
  Testing spring mode switching
  Change some codes for debugging
  Change default mode, add config routes
  Fully functioning wall mode
  Working wall code with parameter setting
  Tie sw1 to motordriver reset
  Enable motor driver reset
  Correct what the PIC prints
  Finish K-setting routine

+++ lib/joy.h
@@ -38,10 +38,10 @@
 void init_joy(void);
 
 typedef struct {
-    uint8_t mode;
+    uint16_t mode;
 
     // spring
-    uint16_t K;
+    float K;
     // wall
     int16_t left;
     int16_t right;
